Add files via upload
Another option is to use Keyboard Maestro with my script which will allow you to use any SINGLE key instead of Apple's Keyboard Shortcuts that require a key combination or a Function key (a Fn key can be a single key). I know some studios like to use the "Enter" key" all the way to the right on a full-size keyboard. Here is the KM macro w/ the script embedded. Remember, in C1Pro you will need to set up the edit keyboard shortcut for "Start/Stop Camera Autofocus". In this case I designated the period "."